Cambios-InscripcionRepositoryImpl

+++ src/main/java/es/fplumara/dam1/talleres/service/impl/UsuarioServiceImpl.java
@@ -115,7 +115,7 @@ public class UsuarioServiceImpl implements UsuarioService {
             inscripcionRepository.deleteById(inscripcion.getId());
         }
 
-        userRepository.deleteById(idUsuario.intValue());
+        userRepository.deleteById((long) idUsuario.intValue());
 
         return new DeleteSummaryDTO(inscripcions.size());
     }
